requirements/base.txt
django<1.12
djangorestframework<3.8
django-crispy-forms<1.8
psycopg2-binary<2.8
python-dateutil<2.8
requests<2.2
